move aggregate_scalar to its own file, make inline
